Split native shell model mapping
