+++ src/gui_runtime/native_shell_runtime/model_mapping/browser.rs
@@ -1,0 +1,180 @@
+use super::*;
+
+impl From<runtime_contract::FocusContextModel> for FocusContextModel {
+    fn from(value: runtime_contract::FocusContextModel) -> Self {
+        match value {
+            runtime_contract::FocusContextModel::None => Self::None,
+            runtime_contract::FocusContextModel::Timeline => Self::Waveform,
+            runtime_contract::FocusContextModel::ContentList => Self::SampleBrowser,
+            runtime_contract::FocusContextModel::NavigationTree => Self::SourceFolders,
+            runtime_contract::FocusContextModel::NavigationList => Self::SourcesList,
+        }
+    }
+}
+
+impl From<FocusContextModel> for runtime_contract::FocusContextModel {
+    fn from(value: FocusContextModel) -> Self {
+        match value {
+            FocusContextModel::None => Self::None,
+            FocusContextModel::Waveform => Self::Timeline,
+            FocusContextModel::SampleBrowser => Self::ContentList,
+            FocusContextModel::SourceFolders => Self::NavigationTree,
+            FocusContextModel::SourcesList => Self::NavigationList,
+        }
+    }
+}
+
+impl From<&SourcesPanelModel> for runtime_contract::SourcesPanelModel {
+    fn from(value: &SourcesPanelModel) -> Self {
+        value.clone()
+    }
+}
+
+impl From<runtime_contract::BrowserPanelModel> for BrowserPanelModel {
+    fn from(value: runtime_contract::BrowserPanelModel) -> Self {
+        Self {
+            visible_count: value.visible_count,
+            selected_visible_row: value.selected_visible_row,
+            autoscroll: value.autoscroll,
+            view_start_row: value.view_start_row,
+            selected_path_count: value.selected_item_count,
+            search_query: value.search_query,
+            active_rating_filters: value.active_rating_filters,
+            active_playback_age_filters: value.active_recency_filters,
+            marked_filter_active: value.marked_filter_active,
+            tag_named_filter_active: value.derived_label_filter_active,
+            tag_named_filter_negated: value.derived_label_filter_negated,
+            sidebar_filters: Default::default(),
+            search_placeholder: value.search_placeholder,
+            busy: value.busy,
+            source_loading: value.data_loading,
+            metadata_pending: value.metadata_pending,
+            file_op_pending: value.mutation_pending,
+            similarity_filtered: value.similarity_filtered,
+            duplicate_cleanup_active: value.duplicate_cleanup_active,
+            sort_label: value.sort_label,
+            active_tab_label: value.active_tab_label,
+            focused_sample_label: value.focused_item_label,
+            tag_sidebar: value.pill_editor,
+            anchor_visible_row: value.anchor_visible_row,
+            rows: retained_vec_from_compat(value.rows),
+        }
+    }
+}
+
+impl From<BrowserPanelModel> for runtime_contract::BrowserPanelModel {
+    fn from(value: BrowserPanelModel) -> Self {
+        Self {
+            visible_count: value.visible_count,
+            selected_visible_row: value.selected_visible_row,
+            autoscroll: value.autoscroll,
+            view_start_row: value.view_start_row,
+            selected_item_count: value.selected_path_count,
+            search_query: value.search_query,
+            active_rating_filters: value.active_rating_filters,
+            active_recency_filters: value.active_playback_age_filters,
+            marked_filter_active: value.marked_filter_active,
+            derived_label_filter_active: value.tag_named_filter_active,
+            derived_label_filter_negated: value.tag_named_filter_negated,
+            search_placeholder: value.search_placeholder,
+            busy: value.busy,
+            data_loading: value.source_loading,
+            metadata_pending: value.metadata_pending,
+            mutation_pending: value.file_op_pending,
+            similarity_filtered: value.similarity_filtered,
+            duplicate_cleanup_active: value.duplicate_cleanup_active,
+            sort_label: value.sort_label,
+            active_tab_label: value.active_tab_label,
+            focused_item_label: value.focused_sample_label,
+            pill_editor: value.tag_sidebar,
+            anchor_visible_row: value.anchor_visible_row,
+            rows: retained_vec_to_compat(value.rows),
+        }
+    }
+}
+
+impl From<&BrowserPanelModel> for runtime_contract::BrowserPanelModel {
+    fn from(value: &BrowserPanelModel) -> Self {
+        value.clone().into()
+    }
+}
+
+impl From<runtime_contract::BrowserChromeModel> for BrowserChromeModel {
+    fn from(value: runtime_contract::BrowserChromeModel) -> Self {
+        Self {
+            samples_tab_label: value.items_tab_label,
+            sample_column_label: value.item_column_label,
+            map_tab_label: value.map_tab_label,
+            tag_editor_label: value.pill_editor_label,
+            search_prefix_label: value.search_prefix_label,
+            search_placeholder: value.search_placeholder,
+            activity_ready_label: value.activity_ready_label,
+            activity_busy_label: value.activity_busy_label,
+            sort_prefix_label: value.sort_prefix_label,
+            sort_order_label: value.sort_order_label,
+            similarity_toggle_label: value.similarity_toggle_label,
+            item_count_label: value.item_count_label,
+        }
+    }
+}
+
+impl From<BrowserChromeModel> for runtime_contract::BrowserChromeModel {
+    fn from(value: BrowserChromeModel) -> Self {
+        Self {
+            items_tab_label: value.samples_tab_label,
+            item_column_label: value.sample_column_label,
+            map_tab_label: value.map_tab_label,
+            pill_editor_label: value.tag_editor_label,
+            search_prefix_label: value.search_prefix_label,
+            search_placeholder: value.search_placeholder,
+            activity_ready_label: value.activity_ready_label,
+            activity_busy_label: value.activity_busy_label,
+            sort_prefix_label: value.sort_prefix_label,
+            sort_order_label: value.sort_order_label,
+            similarity_toggle_label: value.similarity_toggle_label,
+            item_count_label: value.item_count_label,
+        }
+    }
+}
+
+impl From<&BrowserChromeModel> for runtime_contract::BrowserChromeModel {
+    fn from(value: &BrowserChromeModel) -> Self {
+        value.clone().into()
+    }
+}
+
+impl From<runtime_contract::BrowserActionsModel> for BrowserActionsModel {
+    fn from(value: runtime_contract::BrowserActionsModel) -> Self {
+        Self {
+            can_rename: value.can_rename,
+            can_delete: value.can_delete,
+            can_tag: value.can_edit_pills,
+            can_normalize_focused_sample: value.can_process_focused_item,
+            can_loop_crossfade_focused_sample: value.can_open_focused_item_flow,
+            random_navigation_enabled: value.random_navigation_enabled,
+            duplicate_cleanup_active: value.duplicate_cleanup_active,
+            tag_sidebar_open: value.pill_editor_open,
+        }
+    }
+}
+
+impl From<BrowserActionsModel> for runtime_contract::BrowserActionsModel {
+    fn from(value: BrowserActionsModel) -> Self {
+        Self {
+            can_rename: value.can_rename,
+            can_delete: value.can_delete,
+            can_edit_pills: value.can_tag,
+            can_process_focused_item: value.can_normalize_focused_sample,
+            can_open_focused_item_flow: value.can_loop_crossfade_focused_sample,
+            random_navigation_enabled: value.random_navigation_enabled,
+            duplicate_cleanup_active: value.duplicate_cleanup_active,
+            pill_editor_open: value.tag_sidebar_open,
+        }
+    }
+}
+
+impl From<&BrowserActionsModel> for runtime_contract::BrowserActionsModel {
+    fn from(value: &BrowserActionsModel) -> Self {
+        value.clone().into()
+    }
+}
